+++ src/r3/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 **The design has been wholly revamped!** `r3` now only defines the interface between an application and a kernel implementation. The current kernel implementation has been moved to `r3_kernel`. Different kernel implementations that use more exotic architectures (such as interrupt-driven multi-threading) or are built on top on existing RTOSes may be added in the future.
 
+### Added
+
+ - `r3::time::{Duration, Time}: const Default`
+
 ### Changed
 
 While much of the application-level API has retained its general shape, there are some significant changes that may require attention: